% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensorglm.R
\name{glmtf}
\alias{glmtf}
\title{Run GLM using TensorFlow}
\usage{
glmtf(formula, family = gaussian, data, n = 1000, lr = 0.5,
  clip = 1, scale = TRUE, verbose = FALSE)
}
\arguments{
\item{formula}{(like in glm)}

\item{family}{family and link (like in glm)}

\item{data}{data.frmae}

\item{n}{number of GD iterations}

\item{lr}{learning rate}

\item{clip}{clip value}

\item{scale}{divide X variables by max before ajusting?}

\item{verbose}{print model info}
}
\description{
Current implementation works for gaussian, Gamma and Poisson
}
